타입스크립트로 프로젝트 형식 변경을 위해 모듈파일 위치 변경
